<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsScore.R
\name{gsScore}
\alias{gsScore}
\title{Calculate score across genes and samples}
\usage{
gsScore(gm, summarizationFunction = "PC")
}
\arguments{
\item{gm}{normalized count matrix; rows are all genes in the signature that shall be summarized into one score; columns are samples}

\item{summarizationFunction}{character vector defining whether signature scores shall be based on principal component 1 ("PC", default) or z-scores (other value)}
}
\value{
numeric vector of length ncol(gm); a score summarizing the rows of gm
}
\description{
This wrapper function combines filtering out genes with low reads in a number of samples (recommended for limma:voom) with normalization
}
\author{
Dorothee Nickles
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsScore.R
\name{gsScore}
\alias{gsScore}
\title{Calculate score across genes and samples}
\usage{
gsScore(gm, summarizationFunction = "PC")
}
\arguments{
\item{gm}{normalized count matrix; rows are all genes in the signature that shall be summarized into one score; columns are samples}

\item{summarizationFunction}{character vector defining whether signature scores shall be based on principal component 1 ("PC", default) or z-scores (other value)}
}
\value{
numeric vector of length ncol(gm); a score summarizing the rows of gm
}
\description{
This wrapper function combines filtering out genes with low reads in a number of samples (recommended for limma:voom) with normalization
}
\author{
Dorothee Nickles
}
>>>>>>> 5061c1ee8b35a3d7a6a48786e7b3c6eaba7d16b7
