% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_2deltaCt.R
\name{pcr_2deltaCt}
\alias{pcr_2deltaCt}
\title{pcr_2deltaCt}
\usage{
pcr_2deltaCt(
  data,
  Sample = "Sample",
  ref_gene = "ACTB",
  Dup.ID = c("D1", "D2", "D3"),
  path = "0-QC-data"
)
}
\arguments{
\item{data}{Sample is in the first column, next is gene}

\item{Sample}{name of Sample}

\item{ref_gene}{genes as reference}

\item{Dup.ID}{default is c("D1", "D2", "D3"), so the same sample should be stay together}
}
\value{

}
\description{
pcr_2deltaCt
}
\examples{
鍙傝€冮摼鎺ワ細https://zouhua.top/archives/3db8e2e5.html
鍙傝€冮摼鎺ワ細https://www.biolab.wang/r/r_for_qpcr

}
