% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore.R
\name{tmescore}
\alias{tmescore}
\title{calculate TMEscore}
\usage{
tmescore(eset = eset, pdata = pdata, method = method)
}
\arguments{
\item{eset}{normalized expression data use function: scale(); with human gene symbol in rows and patients ID in columns;}

\item{pdata}{phenotype data with patients' ID or other clinical data in columns;the number of rows in pdata should be equal to the number of columns in eset}

\item{method}{use PCA(default) or z-score function to calculate TMEscore}
}
\value{
pdata with TMEscoreA, TMEscoreB and TMEscore
}
\description{
calculate TMEscore
}
\examples{
library('tidyverse')
# system.file("extdata","example.RData",package = "TMEscore")
tmescore<-tmescore(eset = eset,pdata = pdata,method = "pca")

TMEscore = TMEscoreA - TMEscoreB
}
\author{
Dongqiang Zeng
}
