% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_heatmap.R
\name{tmescore_heatmap}
\alias{tmescore_heatmap}
\title{Title}
\usage{
tmescore_heatmap(
  score,
  pcr_data,
  BOR = "BOR",
  patient_ids = NULL,
  batch = NULL,
  width,
  index = 1,
  fig.path = "1-TMEscore-and-genes-heatmap"
)
}
\arguments{
\item{score}{}

\item{pcr_data}{}

\item{BOR}{}

\item{patient_ids}{}

\item{batch}{}

\item{width}{}

\item{index}{}

\item{fig.path}{}
}
\value{

}
\description{
Title
}
