% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_location.R
\name{tmescore_location}
\alias{tmescore_location}
\title{Location of TMEscore of each patient}
\usage{
tmescore_location(
  score,
  vars = c("TMEscore", "TMEscoreA", "TMEscoreB"),
  panel = "ORR",
  method = "method2",
  palette = "nrc",
  palette_line = "jama",
  showplot = TRUE,
  path = NULL,
  tmescore_x = 7,
  tmescore_ab_x = 2,
  tmescore_y_index = -0.5,
  tmescoreab_y_index = 0.5,
  fig.type = "png"
)
}
\arguments{
\item{score}{score}

\item{vars}{default is TMEscore, TMEscroeA and TMEscoreB}

\item{panel}{"PFS" or "ORR"}

\item{method}{default is method2}

\item{palette}{palette of response}

\item{palette_line}{palette of line}

\item{showplot}{default is FALSE}

\item{path}{path to save result}

\item{tmescore_x}{7}

\item{tmescore_ab_x}{default is 3.8}

\item{fig.type}{default is png}
}
\value{

}
\description{
Location of TMEscore of each patient
}
