% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_location3.R
\name{tmescore_location3}
\alias{tmescore_location3}
\title{Location of TMEscore of each patient}
\usage{
tmescore_location3(
  score,
  vars = c("TMEscore", "TMEscoreA", "TMEscoreB"),
  panel = "PFS",
  palette = "nrc",
  palette_line = "jama",
  showplot = TRUE,
  path = NULL,
  x_label = 2.5
)
}
\arguments{
\item{score}{score}

\item{vars}{default is TMEscore, TMEscroeA and TMEscoreB}

\item{panel}{"PFS" or "ORR"}

\item{palette}{palette of response}

\item{palette_line}{palette of line}

\item{showplot}{default is FALSE}

\item{path}{path to save result}

\item{x_label}{default is 2.5}
}
\value{

}
\description{
Location of TMEscore of each patient
}
