% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_pcr.R
\name{tmescore_pcr}
\alias{tmescore_pcr}
\title{TMEscore calculation}
\usage{
tmescore_pcr(
  eset,
  scale = T,
  sig = "pcr",
  method = "mean",
  max = 20,
  mini_gene_count = 2,
  print_gene_pro = T,
  save_eset = F,
  file_name = NULL
)
}
\arguments{
\item{max}{}
}
\value{

}
\description{
TMEscore calculation
}
\author{
Dongqiang Zeng
}
