% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_pcr2.R
\name{tmescore_pcr2}
\alias{tmescore_pcr2}
\title{TMEscore calculation}
\usage{
tmescore_pcr2(
  eset,
  scale = F,
  sig = "pcr",
  max = 20,
  method = "mean",
  mini_gene_count = 2,
  print_gene_pro = T,
  save_eset = F,
  file_name = NULL
)
}
\arguments{
\item{sig}{}
}
\value{

}
\description{
TMEscore calculation
}
\author{
Dongqiang Zeng
}
