% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_value.R
\name{tmescore_value}
\alias{tmescore_value}
\title{Title}
\usage{
tmescore_value(
  score,
  pcr_data,
  BOR = "BOR",
  index = "1-method1",
  jitter = TRUE,
  binwidth = 3,
  addtable = T,
  palette_box = "jama",
  palette_roc = "jama",
  palette_his = "jco",
  fig.type = "png",
  fig.path = "1-testing-score-value"
)
}
\arguments{
\item{addtable}{}
}
\value{

}
\description{
Title
}
