% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_location.R
\name{tmescore_location}
\alias{tmescore_location}
\title{Location of TMEscore of each patient}
\usage{
tmescore_location(
  score,
  vars = c("TMEscore", "TMEscoreA", "TMEscoreB"),
  panel = "ORR",
  method = "method2",
  regimen = "all",
  palette = "nrc",
  palette_line = "jama",
  showplot = TRUE,
  path = NULL,
  tmescore_x = 7,
  tmescore_ab_x = 5.9
)
}
\arguments{
\item{score}{score}

\item{vars}{default is TMEscore, TMEscroeA and TMEscoreB}

\item{panel}{"PFS" or "ORR"}

\item{method}{}

\item{regimen}{}

\item{palette}{palette of response}

\item{palette_line}{palette of line}

\item{showplot}{default is FALSE}

\item{path}{path to save result}

\item{tmescore_x}{7}

\item{tmescore_ab_x}{5.9}
}
\value{

}
\description{
Location of TMEscore of each patient
}
