% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_pcr.R
\name{tmescore_pcr}
\alias{tmescore_pcr}
\title{TMEscore calculation}
\usage{
tmescore_pcr(
  eset,
  scale = F,
  method = "mean",
  mini_gene_count = 2,
  print_gene_pro = T
)
}
\arguments{
\item{mini_gene_count}{}
}
\value{

}
\description{
TMEscore calculation
}
\author{
Dongqiang Zeng
}
