% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmescore_pcr4.R
\name{tmescore_pcr4}
\alias{tmescore_pcr4}
\title{TMEscore calculation}
\usage{
tmescore_pcr4(
  eset,
  scale = F,
  sig = "pcr",
  max = 20,
  method = "mean",
  mini_gene_count = 2,
  print_gene_pro = T,
  coef = 1,
  save_eset = F,
  file_name = NULL
)
}
\arguments{
\item{max}{}
}
\value{

}
\description{
TMEscore calculation
}
\author{
Dongqiang Zeng
}
